+++ doc/source/history.rst
@@ -3,6 +3,7 @@ Historique
 
 Versions:
 
+* 0.9.2 : Factorisation du code - disparition des fonctions internes en _simplif, ajout d'un paramètre simplif (True/False)
 * 0.9 : Ajout de la fonction *randExpr* pour générer aléatoirement des produits/quotients.
         Une fonction annexe *_degree_good* ajoutée suite à un défaut sur la fonction degree de sympy (qui lève une exception quand on lui donne une constante).
 	Retrait de la classe TableauVariation qui n'est pas opérationnelle